Convert back to modules, still use window global
