+++ demo/pages/_document.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { Html, Head, Main, NextScript } from 'next/document';
-import Script from 'next/script';
 
 const Document: React.FC = () => {
   return (
@@ -9,7 +8,6 @@ const Document: React.FC = () => {
       <body>
         <Main />
         <NextScript />
-        <Script src="sdk/index.js" strategy="beforeInteractive"></Script>
       </body>
     </Html>
   );
