+++ tsconfig.json
@@ -3,9 +3,9 @@
     "declaration": true,
     "inlineSources": true,
     "lib": ["ES6", "DOM"],
-    "module": "none",
+    "module": "commonjs",
     "esModuleInterop": true,
-    "outFile": "dist/index.js",
+    "outDir": "dist",
     "rootDir": "src",
     "sourceMap": true,
     "strict": true,
